backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
langchain==0.3.3
langchain-openai==0.2.0
langgraph==0.2.34
openai==1.51.2
redis==5.0.8
alembic==1.13.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
mem0ai
litellm

